Bounds check API update + Metadata aritrary bytes (#282)
Update types' APIs that have bounds restrictions e.g. metadata
text/bytes, crypto primitives, etc with new codegen update.
Now these types (that aren't newtypes - this was always supported) will
have checks on the constructors/setters, not just in deserialization.

Moved over the arbitrary byte chunking utils for metadatums from old CML
to new.

+++ chain/rust/src/assets/mod.rs
@@ -38,7 +38,7 @@ impl AssetName {
             return Err(DeserializeError::new(
                 "AssetName",
                 DeserializeFailure::RangeCheck {
-                    found: inner.len(),
+                    found: inner.len() as isize,
                     min: Some(0),
                     max: Some(32),
                 },
